added an authentication , user model, an admin ,generate forgot password function

+++ controllers/bootcamps.js
@@ -31,6 +31,22 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 //@route      POST /api/v1/bootcamps
 //@access     Private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
+  //Add user to req.body
+  req.body.user = req.user.id;
+
+  //check for published bootcamp
+  const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id });
+
+  //If the user is not an admin, they can only add one bootcamp
+  if (publishedBootcamp && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `The user with ID ${req.user.id} has already published a bootcamp`,
+        400
+      )
+    );
+  }
+
   const bootcamp = await Bootcamp.create(req.body);
 
   res.status(201).json({
@@ -43,16 +59,25 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 //@route      PUT /api/v1/bootcamps/:id
 //@access     Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  let bootcamp = await Bootcamp.findById(req.params.id);
 
   if (!bootcamps) {
     return next(
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
     );
   }
+
+  // Make sure user is bootcamp owner
+  if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401)
+    );
+  }
+
+  bootcamp = await Bootcamp.findOneAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
+  });
   res.status(200).json({ success: true, data: bootcamp });
 });
 
@@ -69,6 +94,13 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
     );
   }
+
+  // Make sure user is bootcamp owner
+  if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorized to delete this bootcamp`, 401)
+    );
+  };
 
   // Delete related courses
   await course.deleteMany({ bootcamp: req.params.id });
@@ -124,34 +156,46 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
     );
   }
 
+  // Make sure user is bootcamp owner
+  if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+    return next(
+       new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401)
+    );
+  };
+
   // Check if file is uploaded
-  if(!req.files){
+  if (!req.files) {
     return next(new ErrorResponse(`Please upload a file`, 400));
   }
 
   const file = req.files.file;
 
   // Check if the file is an image
-  if(!file.mimetype.startsWith('image')){
+  if (!file.mimetype.startsWith("image")) {
     return next(new ErrorResponse(`Please upload an image file`, 400));
   }
 
   // Check file size
-  if(file.size > process.eventNames.MAX_FILE_UPLOAD){
-    return next(new ErrorResponse(`Please upload an image less than ${process.eventNames.MAX_FILE_UPLOAD}`, 400));
+  if (file.size > process.eventNames.MAX_FILE_UPLOAD) {
+    return next(
+      new ErrorResponse(
+        `Please upload an image less than ${process.eventNames.MAX_FILE_UPLOAD}`,
+        400
+      )
+    );
   }
 
   //create custom filename
   file.name = `photo_${bootcamp._id}${path.parse(file.name).ext}`;
 
-  file.mv(`${process.env.FILE_UPLOAD_PATH}/${file.name}`, async err => {
-    if(err){
+  file.mv(`${process.env.FILE_UPLOAD_PATH}/${file.name}`, async (err) => {
+    if (err) {
       console.error(err);
       return next(new ErrorResponse(`Problem with file upload`, 500));
     }
 
-    await Bootcamp.findByIdAndUpdate(req.params.id, {photo: file.name});
-  })
+    await Bootcamp.findByIdAndUpdate(req.params.id, { photo: file.name });
+  });
 
   // Send response
   res.status(200).json({
